refeat: forms layout and new routes
