+++ src/components/FormsComissao/cadastrar/CadastroComissao.tsx
@@ -1,16 +1,39 @@
 'use client';
 
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 
+import axios from 'axios';
+import { result } from 'lodash';
+import { FaTimes } from 'react-icons/fa';
 import { FiEye, FiEyeOff } from 'react-icons/fi';
+import Select from 'react-select';
 
-export default function CadastroConvidado() {
+import AlertCard from '@/components/AlertCard';
+import { Area } from '@/lib/repository/area/index.repository';
+import { Comissao } from '@/lib/repository/comission/index.repository';
+import mockedOptionAreas from '@/mocks/OptionsAreas';
+import mockedOptionTurnos from '@/mocks/OptionsTurnos';
+
+type LabelValue = {
+	label: string;
+	value: string | undefined;
+};
+
+export default function CadastroComissao() {
 	const [passwordVisible, setPasswordVisible] = useState(false);
 	const [password, setPassword] = useState('');
+	const [name, setName] = useState('');
+	const [cpf, setCpf] = useState('');
+	const [email, setEmail] = useState('');
+	const [instituicao, setInst] = useState('');
+	const [turno, setTurno] = useState<string | undefined>('');
+	const [lattes, setLattes] = useState('');
 	const [confirmpasswordVisible, setConfirmpasswordVisible] = useState(false);
 	const [confirmpassword, setConfirmpassword] = useState('');
-	const checkboxNames = ['Palestrante', 'Mentor', 'Outro'];
+	// funcao no evento:
+	const checkboxNames = ['Organizador', 'Chair', 'Avaliador', 'Admin'];
 	const [checkboxes, setCheckboxes] = useState(checkboxNames.map(() => false));
+	const [showCard, setShowCard] = useState(false);
 
 	const handleCheckboxChange = (index: any) => {
 		setCheckboxes((prev) => {
@@ -20,11 +43,112 @@ export default function CadastroConvidado() {
 		});
 	};
 
+	const [areas, setAreas] = useState<LabelValue[]>([
+		{ value: '1', label: 'Area 1' },
+		{ value: '2', label: 'Area 2' },
+	]);
+	const [realAreas, setRealAreas] = useState<(string | undefined)[]>([]);
+
+	const customStyles = {
+		control: (provided: any) => ({
+			...provided,
+			width: '100%',
+			height: 'auto',
+			borderRadius: '0.375rem',
+			border: '1',
+			background: 'white',
+			fontSize: '0.875rem',
+		}),
+	};
+
 	const handleTogglePasswordVisibility = () => {
 		setPasswordVisible(!passwordVisible);
 	};
 	const handleToggleConfirmPasswordVisibility = () => {
 		setConfirmpasswordVisible(!confirmpasswordVisible);
+	};
+
+	useEffect(() => {
+		async function getAreas() {
+			try {
+				const id = localStorage.getItem('eventId');
+
+				const result = await axios.get(
+					`http://localhost:5002/area-event/${id}`
+				);
+				if (result.data.areas) {
+					const areasComming: Area[] = result.data.areas;
+					const areasValueLabel = areasComming.map((area) => {
+						const labelvalue = { value: area.id, label: area.nome };
+						return labelvalue;
+					});
+					console.log(areasValueLabel);
+					setAreas(areasValueLabel);
+				}
+			} catch (error) {
+				console.log(error);
+			}
+		}
+		getAreas();
+	}, []);
+
+	const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
+		e.preventDefault();
+
+		try {
+			const organizador = checkboxes.filter((item) =>
+				item === checkboxes[0] ? true : false
+			)[0];
+			const adm = checkboxes.filter((item) =>
+				item == checkboxes[3] ? true : false
+			)[0];
+			const chair = checkboxes.filter((item) =>
+				item == checkboxes[1] ? true : false
+			)[0];
+			const avaliador = checkboxes.filter((item) =>
+				item == checkboxes[2] ? true : false
+			)[0];
+			console.log(realAreas?.map((area) => area));
+			const data: Comissao = {
+				name,
+				email,
+				senha: password,
+				cpf,
+				instituicao,
+				turno,
+				lattes,
+				// adm: checkboxes[3],
+				// organizador: checkboxes[0],
+				// chair: checkboxes[1],
+				// avaliador: checkboxes[2],
+				adm,
+				organizador,
+				chair,
+				avaliador,
+				// certificado: '',
+				areaConhecimento: realAreas?.map((area) => area),
+			};
+			const result = await axios.post('http://localhost:5002/comissao', data);
+			console.log(result);
+			if (result.data.comissaoCreated) {
+				// habilitar card de 3seg indicando cadastro realizado
+				console.log(result);
+				setShowCard(true);
+				setTimeout(() => {
+					setShowCard(false);
+				}, 3000);
+				setName('');
+				setCpf('');
+				setEmail('');
+				setInst('');
+				setPassword('');
+				setConfirmpassword('');
+				setLattes('');
+				setCheckboxes(checkboxNames.map(() => false));
+			}
+		} catch (error) {
+			console.log(error);
+		}
 	};
 
 	return (
@@ -34,12 +158,13 @@ export default function CadastroConvidado() {
 					className="text-center text-2xl font-bold text-black"
 					style={{ color: '#4B00E0' }}
 				>
-					Cadastrar Convidado
+					Cadastrar Comissão
 				</h1>
+				<AlertCard message="Comissao cadastrada com sucesso" show={showCard} />
 				<p className="text-center text-sm text-black">
-					Cadastro como convidado
+					Cadastro como parte da comissão, possível mais de uma função
 				</p>
-				<form className="mt-8 w-full card">
+				<form className="card mt-8" onSubmit={handleSubmit}>
 					<div className="flex justify-center gap-5">
 						<div className="w-full">
 							<div className="mb-5 flex flex-col">
@@ -53,7 +178,9 @@ export default function CadastroConvidado() {
 										type="text"
 										name="fullName"
 										id="fullName"
-										placeholder="Nome do Convidado"
+										placeholder="Nome da Comissão"
+										value={name}
+										onChange={(e) => setName(e.target.value)}
 										required
 									/>
 								</div>
@@ -69,7 +196,9 @@ export default function CadastroConvidado() {
 										type="text"
 										name="cpf"
 										id="cpf"
-										placeholder="CPF do Convidado"
+										placeholder="CPF da Comissão"
+										value={cpf}
+										onChange={(e) => setCpf(e.target.value)}
 										required
 									/>
 								</div>
@@ -85,13 +214,13 @@ export default function CadastroConvidado() {
 										type="email"
 										name="email"
 										id="email"
-										placeholder="emailconvidado@email.com"
+										placeholder="emailcomissao@email.com"
+										value={email}
+										onChange={(e) => setEmail(e.target.value)}
 										required
 									/>
 								</div>
 							</div>
-						</div>
-						<div className="w-full">
 							<div className="mb-5 flex flex-col">
 								<label className="mb-2 text-sm font-medium" htmlFor="password">
 									Senha
@@ -104,7 +233,7 @@ export default function CadastroConvidado() {
 										onChange={(e) => setPassword(e.target.value)}
 										name="password"
 										id="password"
-										placeholder="Senha do Convidado"
+										placeholder="Senha da Comissão"
 										required
 									/>
 									{passwordVisible ? (
@@ -132,7 +261,7 @@ export default function CadastroConvidado() {
 										onChange={(e) => setConfirmpassword(e.target.value)}
 										name="password"
 										id="password"
-										placeholder="Senha do Convidado"
+										placeholder="Senha da Comissão"
 										required
 									/>
 									{confirmpasswordVisible ? (
@@ -148,7 +277,81 @@ export default function CadastroConvidado() {
 									)}
 								</div>
 							</div>
-							<div className="mb-4">
+						</div>
+						<div className="w-full">
+							<div className="mb-5 flex flex-col">
+								<label className="mb-2 text-sm font-medium" htmlFor="areas">
+									Áreas de Conhecimento
+								</label>
+								<div className="w-full">
+									<Select
+										isMulti
+										name="areas"
+										options={areas}
+										className="basic-multi-select border-gray-300"
+										classNamePrefix="select"
+										styles={customStyles}
+										onChange={(choice) =>
+											setRealAreas(choice.map((a) => a.value))
+										}
+									/>
+								</div>
+							</div>
+							<div className="mb-5 flex flex-col">
+								<label
+									className="mb-2 text-sm font-medium"
+									htmlFor="instituicao"
+								>
+									Instituição Referente
+								</label>
+
+								<div className="rounded-md border border-gray-300 bg-white px-4 py-2">
+									<input
+										className="w-full rounded-md border-0 bg-white text-sm outline-none"
+										type="text"
+										name="instituicao"
+										id="instituicao"
+										placeholder="Instituição da Comissão"
+										value={instituicao}
+										onChange={(e) => setInst(e.target.value)}
+										required
+									/>
+								</div>
+							</div>
+							<div className="mb-5 flex flex-col">
+								<label className="mb-2 text-sm font-medium" htmlFor="turno">
+									Turno
+								</label>
+								<div className="w-full">
+									<Select
+										name="turnos"
+										options={mockedOptionTurnos}
+										className="basic-multi-select border-gray-300"
+										classNamePrefix="select"
+										styles={customStyles}
+										onChange={(choice) => setTurno(choice?.label)}
+									/>
+								</div>
+							</div>
+							<div className="mb-5 flex flex-col">
+								<label className="mb-2 text-sm font-medium" htmlFor="link">
+									Link Lattes
+								</label>
+
+								<div className="rounded-md border border-gray-300 bg-white px-4 py-2">
+									<input
+										className="w-full rounded-md border-0 bg-white text-sm outline-none"
+										type="url"
+										name="link"
+										id="link"
+										placeholder="link Lattes da Comissão"
+										value={lattes}
+										onChange={(e) => setLattes(e.target.value)}
+										required
+									/>
+								</div>
+							</div>
+							<div className="mb-5">
 								<label className="mb-2 text-sm font-medium" htmlFor="funcao">
 									Função no Evento
 								</label>
@@ -212,7 +415,10 @@ export default function CadastroConvidado() {
 					<div className="mb-6">
 						<p className="text-center text-xs font-normal text-slate-400">
 							Já tem uma conta？
-							<a className="font-bold underline text-[#4B00E0] cursor-pointer" href='/login'>
+							<a
+								className="cursor-pointer font-bold text-[#4B00E0] underline"
+								href="/login"
+							>
 								Log in
 							</a>
 						</p>
@@ -220,7 +426,7 @@ export default function CadastroConvidado() {
 					<div className="flex items-center justify-center gap-5">
 						<button
 							className="mb-6 w-1/5 rounded-xl border-none p-2 text-center text-base font-medium text-white"
-							type="submit"
+							type="button"
 							style={{ backgroundColor: '#8A8A8A' }}
 						>
 							Voltar
